# src/backend/model/user.dto.ts
import { IsString, IsOptional } from 'class-validator'
import { ApiProperty } from '@nestjs/swagger'

export class UserDto {
  @IsString()
  @IsOptional()
  @ApiProperty()
  id: string;

  @IsString()
  @ApiProperty()
  first_name: string;

  @IsString()
  @ApiProperty()
  last_name: string;

  @IsString()
  @ApiProperty()
  gender: string;
}